Simplify server by using gRPC error system
