+++ clients/jvm/src/main/kotlin/io/cloudevents/protobuf/adapter.kt
@@ -4,10 +4,8 @@ import io.cloudevents.CloudEvent
 import io.cloudevents.core.builder.CloudEventBuilder.fromSpecVersion
 import io.cloudevents.v1.proto.CloudEvent as ProtoCloudEvent
 
-// ***** CloudEvents themselves (the envelope) *****
-
 fun ProtoCloudEvent.toDomain(): CloudEvent =
     ProtoDeserializer(this).read(::fromSpecVersion)
 
-fun CloudEvent.toProto(): ProtoCloudEvent =
+fun CloudEvent.toTransfer(): ProtoCloudEvent =
     ProtoSerializer.toProto(this)
