+++ clients/settings.gradle.kts
@@ -12,16 +12,17 @@ dependencyResolutionManagement {
 
     versionCatalogs {
         create("libs") {
+            version("arrow", "1.2.1")
             version("caffeine", "3.1.1")
             version("cloudevents", "2.5.0")
-            version("protobuf-java", "3.23.3")
+            version("protobuf-java", "3.25.1")
             version("grpc", "1.56.0")
             version("grpckotlin", "1.3.0")
             version("slf4j", "1.7.36")
 
             library("kotlinx-coroutines", "org.jetbrains.kotlinx", "kotlinx-coroutines-core").version("1.7.1")
             library("kotlinx-coroutines-jdk8", "org.jetbrains.kotlinx", "kotlinx-coroutines-jdk8").version("1.7.1")
-            library("arrow-core", "io.arrow-kt", "arrow-core").version("1.1.5")
+            library("arrow-core", "io.arrow-kt", "arrow-core").versionRef("arrow")
             library("caffeine", "com.github.ben-manes.caffeine", "caffeine").versionRef("caffeine")
             library("cloudevents-core", "io.cloudevents", "cloudevents-core").versionRef("cloudevents")
             library("cloudevents-protobuf", "io.cloudevents", "cloudevents-protobuf").versionRef("cloudevents")
